requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
celery==5.3.1
certifi==2022.12.7
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
distlib==0.3.6
dj-database-url==1.3.0
Django==4.2.4
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.2
django-ninja==0.22.2
django-phonenumber-field==7.1.0
django-sendgrid==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
elastic-transport==8.12.0
elasticsearch==8.7.0
filelock==3.9.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
kombu==5.3.1
MarkupSafe==2.1.3
numpy==1.24.2
package-name==0.1
packaging==23.1
pandas==1.5.3
phonenumbers==8.13.13
Pillow==10.0.0
platformdirs==3.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.17
pydantic==1.10.12
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
pywin32==305
PyYAML==6.0.1
redis==5.0.0
service-identity==24.1.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.1
vine==5.0.0
virtualenv==20.19.0
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.7
wcwidth==0.2.6
whitenoise==6.5.0
zope.interface==6.2